requirements.txt
streamlit~=1.42.2
mysql-connector-python~=9.2.0